+++ ios/Podfile
@@ -34,6 +34,8 @@ target 'ProjectStructurePoc' do
   pod 'glog', :podspec => '../node_modules/react-native/third-party-podspecs/glog.podspec'
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 
+  pod 'RNReanimated', :path => '../node_modules/react-native-reanimated'
+
   target 'ProjectStructurePocTests' do
     inherit! :search_paths
     # Pods for testing
