+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ProjectStructurePoc",
+  "name": "Project Structure Poc",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -11,7 +11,12 @@
   },
   "dependencies": {
     "react": "16.9.0",
-    "react-native": "0.61.5"
+    "react-native": "0.61.5",
+    "react-native-gesture-handler": "^1.5.3",
+    "react-native-reanimated": "^1.4.0",
+    "react-native-screens": "^2.0.0-alpha.22",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3"
   },
   "devDependencies": {
     "@babel/core": "7.7.7",
